+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to Pekora will be documented here. Breaking changes are mark
 
 Pekora adheres to [semantic versioning](https://semver.org/spec/v2.0.0.html).
 
+## <a name="1-0-2">[1.0.2] - 2023-03-13</a>
+
+### Fixed
+
+- Fixed a bug where subtracting permissions could return an incorrect result if the difference of the permissions'
+  integer values was a negative number.
+
 ## <a name="1-0-1">[1.0.1] - 2023-03-12</a>
 
 No user-facing changes are introduced in this release.